--- src/main/java/roomescape/repository/ReservationRepositoryImpl.java
@@ -1,48 +1,0 @@
-package roomescape.repository;
-
-import java.time.LocalDate;
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.stereotype.Repository;
-
-import lombok.RequiredArgsConstructor;
-import roomescape.domain.Reservation;
-import roomescape.repository.dao.ReservationDao;
-
-@Repository
-@RequiredArgsConstructor
-public class ReservationRepositoryImpl implements ReservationRepository {
-
-    private final ReservationDao reservationDao;
-
-    @Override
-    public List<Reservation> getAll() {
-        return reservationDao.selectAll();
-    }
-
-    @Override
-    public Reservation save(Reservation reservation) {
-        return reservationDao.insertAndGet(reservation);
-    }
-
-    @Override
-    public Optional<Reservation> findById(Long reservationId) {
-        return reservationDao.selectById(reservationId);
-    }
-
-    @Override
-    public Reservation getById(Long id) {
-        return findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 예약입니다."));
-    }
-
-    @Override
-    public void remove(Reservation reservation) {
-        reservationDao.deleteById(reservation.id());
-    }
-
-    @Override
-    public boolean existDuplicatedDateTime(LocalDate date, Long timeId, Long themeId) {
-        return reservationDao.existDuplicatedDateTime(date, timeId, themeId);
-    }
-}
