--- src/main/java/roomescape/repository/ReservationThemeRepositoryImpl.java
@@ -1,47 +1,0 @@
-package roomescape.repository;
-
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.stereotype.Repository;
-
-import lombok.RequiredArgsConstructor;
-import roomescape.domain.ReservationTheme;
-import roomescape.repository.dao.ReservationThemeDao;
-
-@Repository
-@RequiredArgsConstructor
-public class ReservationThemeRepositoryImpl implements ReservationThemeRepository {
-
-    private final ReservationThemeDao reservationThemeDao;
-
-    @Override
-    public List<ReservationTheme> getAll() {
-        return reservationThemeDao.selectAll();
-    }
-
-    @Override
-    public ReservationTheme save(ReservationTheme reservationTheme) {
-        return reservationThemeDao.insertAndGet(reservationTheme);
-    }
-
-    @Override
-    public Optional<ReservationTheme> findById(Long id) {
-        return reservationThemeDao.selectById(id);
-    }
-
-    @Override
-    public ReservationTheme getById(Long id) {
-        return findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 테마입니다."));
-    }
-
-    @Override
-    public void remove(ReservationTheme reservationTheme) {
-        reservationThemeDao.deleteById(reservationTheme.id());
-    }
-
-    @Override
-    public List<ReservationTheme> orderByThemeBookedCountWithLimit(int limit) {
-        return reservationThemeDao.orderByThemeBookedCountWithLimit(limit);
-    }
-}
