+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -21,9 +21,6 @@ import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTheme;
 import roomescape.domain.ReservationTime;
 import roomescape.exception.DBFKException;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ReservationThemeRepository;
-import roomescape.repository.ReservationTimeRepository;
 import roomescape.service.dto.request.CreateReservationTimeServiceRequest;
 import roomescape.service.dto.response.ReservationTimeServiceResponse;
